Added slider, text, scripts, change hierarchy. Removed unnecessary joystick script.
Added sprites for the slider: sliderEmpty.png and sliderFill.png
"Power" Slider. Automatically filled in and indicates the strength of the throw. Also added Russian text above the "Сила броска" slider.
Removed unnecessary joystick script.
Added SliderController.cs script which automatically fills the slider.
Added SnowballButton.cs script that adds actions when clicking on the SnowballButton. Sets the direction, strength and lifespan of an attacking ability.
Added an empty object to the scene, HippoSnowballSet, which contains a certain stock of offensive abilities.