Changed tiles. The level now has depth. Changes to the conditions of the throw. And other.
All tiles have been completely replaced and now everything looks great.
The level has depth. The further away the character, the smaller it is.
Changes to the conditions of the throw. The strength scale after filling smoothly decreases to zero. When thrown, the value of the power scale becomes zero and does not change until a full recharge. The TimerRoundController.cs script code has been split. The code responsible for converting text to a picture is now a separate class in the TextPictureConverter.cs script
Added CloudsController.cs script that moves clouds.
Added status bar to recharge attack ability. When fully recharged, the particle system above the character is activated.